extracted difference box to another file
