+++ src/pages/stake/index.tsx
@@ -64,6 +64,7 @@ import ShadowDiv from './atoms/ShadowDiv'
 import InfoPair from './molecules/InfoPair'
 import CardSectionValue from './components/CardSectionValue'
 import VerticalSeparator from './components/VerticalSeparator'
+import DifferenceNotification from './organisms/DifferenceNotification'
 
 // disable no unused variables
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -343,105 +344,6 @@ function Stake1(): any {
  Components
  */
 
-const DifferenceText = function DifferenceText({
-  children,
-  onClick,
-}: {
-  onClick: () => void
-  children: React.ReactNode
-}) {
-  // text-sm font-bold underline mt-3 text-underline-offset[4px] text-decoration-thickness[2px] self-center cursor-pointer select-none hover:opacity-80 duration-200
-  const StyledText = styled.div`
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    margin-top: 0.75rem;
-    font-weight: 700;
-    text-decoration: underline;
-    text-decoration-thickness: 2px;
-    text-underline-offset: 4px;
-    align-self: center;
-    cursor: pointer;
-    user-select: none;
-    transition: opacity 0.2s ease-in-out;
-    &:hover {
-      opacity: 0.8;
-    }
-  `
-  return (
-    <StyledText
-      onClick={onClick}
-      // className="text-sm font-bold underline mt-3 text-underline-offset[4px] text-decoration-thickness[2px] self-center cursor-pointer select-none hover:opacity-80 duration-200"
-    >
-      {children}
-    </StyledText>
-  )
-}
-// const Notification = tw.div`bg-[#F04D42] text-[#fafafa] rounded-[10px] p-6 text-sm font-medium flex items-center`
-const Notification = styled.div<GeneralElementProps>`
-  background-color: #f04d42;
-  color: #fafafa;
-  padding: 1.5rem;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-  border-radius: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`
-
-const baseButtonStyle = css`
-  border-radius: 0.5rem;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  user-select: none;
-  border: 1px solid;
-  border-color: white;
-`
-const redButtonStyle = css`
-  background-color: #f04d42;
-  color: #fafafa;
-  transition: all 0.2s ease-in-out;
-  cursor: pointer;
-  &:hover {
-    background-color: white;
-    color: #f04d42;
-  }
-`
-const whiteButtonStyle = css`
-  background-color: white;
-  color: #f04d42;
-`
-
-const NotificationButton = styled.div<{ active?: boolean }>`
-  ${baseButtonStyle}
-  ${({ active }) => (active ? whiteButtonStyle : redButtonStyle)}
-  &:not(:first-child) {
-    margin-left: 10px;
-  }
-  height: 34px;
-  width: 117px;
-  border-radius: 10px;
-  font-size: 14px;
-`
-
-const Typography = {
-  Notice: styled.p`
-    margin-top: 0;
-    margin-bottom: 0;
-    margin-right: 60px;
-    font-size: 0.875rem /* 14px */;
-    line-height: 22.4px;
-    font-weight: 500;
-  `,
-  Emphasis: styled.span`
-    font-weight: 700;
-  `,
-} as const
-
 export default function Stake(): JSX.Element {
   // account usewallet
   const { account } = useWallet()
@@ -522,39 +424,8 @@ export default function Stake(): JSX.Element {
         </HeroContainer>
       ) : (
         <Content>
-          <Notification>
-            <Typography.Notice>
-              We have updated our staking mechanism to a new version{' '}
-              <Typography.Emphasis>STAKING V2</Typography.Emphasis> which is a part of our{' '}
-              <Typography.Emphasis>Governance system</Typography.Emphasis>. New staking is available only in new{' '}
-              <Typography.Emphasis>STAKING V2</Typography.Emphasis>. In{' '}
-              <Typography.Emphasis>STAKING V1</Typography.Emphasis> you can unstake your funds or migrate funds to new{' '}
-              <Typography.Emphasis>STAKING V2</Typography.Emphasis>.
-            </Typography.Notice>
-            <div
-              style={{
-                display: 'flex',
-                flexDirection: 'column',
-              }}
-            >
-              <div style={{ display: 'flex' }}>
-                {/* div with 2 buttons horizontally saying Staking V1 and Staking V2, one border white, red bg, white text, the other white bg, red text, both semibold */}
-                <NotificationButton active={version === Version.v1} onClick={() => setVersion(Version.v1)}>
-                  Staking V1
-                </NotificationButton>
-                <NotificationButton
-                  onClick={() => {
-                    setVersion(Version.v2)
-                  }}
-                  active={version === Version.v2}
-                >
-                  Staking V2
-                </NotificationButton>
-              </div>
-              <DifferenceText onClick={() => setVersion(Version.difference)}>What is the difference?</DifferenceText>
-            </div>
-          </Notification>
-          <Checkbox type="checkbox" />
+          <DifferenceNotification version={version} setVersion={setVersion} />
+          {/* <Checkbox type="checkbox" /> */}
           {/* 24-padding white box with 10px radius corner and shadow */}
           <ShadowDiv style={{ marginBottom: '20px' }}>
             <RaisedBox
